+++ esbuild.js
@@ -1,3 +1,4 @@
+const fs = require('node:fs');
 const { build } = require('esbuild');
 
 const baseConfig = {
@@ -19,22 +20,9 @@ const webviewsConfig = {
     ...baseConfig,
     target: 'es2020',
     format: 'esm',
-    entryPoints: [
-        './src/webviews/bucket-details.tsx',
-        './src/webviews/object-details.tsx',
-        './src/webviews/custom-translation.tsx',
-        './src/webviews/thumbnails.tsx',
-        './src/webviews/webhook-details.tsx',
-        './src/webviews/create-webhook.tsx',
-        './src/webviews/update-webhook.tsx',
-        './src/webviews/derivative-preview.tsx',
-        './src/webviews/appbundle-details.tsx',
-        './src/webviews/alias-details.tsx',
-        './src/webviews/activity-details.tsx',
-        './src/webviews/create-activity.tsx',
-        './src/webviews/update-activity.tsx',
-        './src/webviews/create-workitem.tsx'
-    ],
+    entryPoints: fs.readdirSync('./src/webviews')
+        .filter(file => file.endsWith('.tsx'))
+        .map(file => `./src/webviews/${file}`),
     outdir: './out/webviews',
     splitting: true
 };