+++ src/clients/index.ts
@@ -1,0 +1,4 @@
+export { DefaultRequestAdapter } from "./DefaultRequestAdapter.js";
+export { BearerTokenAuthenticationProvider } from "./BearerTokenAuthenticationProvider.js";
+export { ClientCredentialsAuthenticationProvider } from "./ClientCredentialsAuthenticationProvider.js";
+export { createSecureServiceAccountsClient } from "./secure-service-accounts/secureServiceAccountsClient.js";