+++ src/common.ts
@@ -13,8 +13,9 @@ import {
 } from 'aps-sdk-node';
 import { IDerivative } from './interfaces/model-derivative';
 import { IAuthOptions } from 'aps-sdk-node/dist/common';
-import { IEnvironment } from './environments';
+import { IEnvironment } from './environment';
 import { ModelDerivativeFormats, isViewableFormat } from './providers/model-derivative';
+import { createSecureServiceAccountsClient } from './clients';
 
 export interface IPreviewSettings {
     extensions: string[];
@@ -33,6 +34,7 @@ export interface IContext {
     designAutomationClient: DesignAutomationClient;
     webhookClient: WebhooksClient;
     bim360Client: BIM360Client;
+    secureServiceAccountsClient: ReturnType<typeof createSecureServiceAccountsClient>;
     previewSettings: IPreviewSettings;
     threeLeggedToken?: string;
     log: vscode.LogOutputChannel;
@@ -139,6 +141,8 @@ export async function showErrorMessage(title: string, err: any, context?: IConte
     } else if (typeof err === 'object') {
         if (err.message) {
             msg += ': ' + err.message;
+        } else if (err.detail) {
+            msg += ': ' + err.detail;
         }
     }