+++ src/providers/hubs.ts
@@ -62,32 +62,64 @@ export class HubsDataProvider implements vscode.TreeDataProvider<HubsEntry> {
         if (isHub(entry)) {
             const node = new vscode.TreeItem(entry.name, vscode.TreeItemCollapsibleState.Collapsed);
             node.id = entry.id;
+            node.tooltip = [
+                `Hub`,
+                `ID: ${entry.id}`,
+                `Name: ${entry.name}`
+            ].join('\n');
             node.contextValue = 'hub';
             return node;
         } else if (isProject(entry)) {
             const node = new vscode.TreeItem(entry.name, vscode.TreeItemCollapsibleState.Collapsed);
             node.id = entry.id;
+            node.tooltip = [
+                `Project`,
+                `ID: ${entry.id}`,
+                `Name: ${entry.name}`
+            ].join('\n');
             node.contextValue = 'project';
             return node;
         } else if (isFolder(entry)) {
             const node = new vscode.TreeItem(entry.name, vscode.TreeItemCollapsibleState.Collapsed);
             node.id = entry.id;
+            node.tooltip = [
+                `Folder`,
+                `ID: ${entry.id}`,
+                `Name: ${entry.name}`
+            ].join('\n');
             node.contextValue = 'folder';
             return node;
         } else if (isItem(entry)) {
             const node = new vscode.TreeItem(entry.name, vscode.TreeItemCollapsibleState.Collapsed);
             node.id = entry.id;
+            node.tooltip = [
+                `Item`,
+                `ID: ${entry.id}`,
+                `Name: ${entry.name}`
+            ].join('\n');
             node.contextValue = 'item';
             return node;
         } else if (isVersion(entry)) {
             const node = new vscode.TreeItem(entry.name, vscode.TreeItemCollapsibleState.Collapsed);
             node.id = entry.id;
+            node.tooltip = [
+                `Version`,
+                `ID: ${entry.id}`,
+                `Name: ${entry.name}`
+            ].join('\n');
             node.contextValue = 'version';
             return node;
         } else if (isDerivative(entry)) {
             const node = new vscode.TreeItem(entry.name, vscode.TreeItemCollapsibleState.None);
             node.id = entry.urn;
-            node.contextValue = 'derivative';
+            node.tooltip = [
+                `Derivative`,
+                `Name: ${entry.name}`,
+                `Format: ${entry.format}`,
+                `Role: ${entry.role}`
+            ].join('\n');
+            node.iconPath = new vscode.ThemeIcon('file-binary');
+            node.contextValue = entry.nonViewable ? 'non-viewable-derivative' : 'derivative';
             return node;
         } else {
             const node = new vscode.TreeItem('', vscode.TreeItemCollapsibleState.None);