+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+- Added
+  - Support for Secure Service Accounts API
+  - Confirmation modal before deleting any resources
+- Updated
+  - Tooltips on all tree view elements
+
 ## [2.8.0] - 2025-01-21
 
 - Added